+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
+## Unreleased
+
+### Added
+
+* Added --loglevel CLI argument for `run_action_based_chat_baseline.py` script
+
+### Changed
+
+### Fixed
+
+* Fixed issue where TA3 training session flag wasn't being passed to the TA3 API
+
+* Removing training session data info from "action to take" passed to TA3 API
+
+
 ## Version 0.2.2
 
 #### Added